backend/requirements.txt
aiohttp==3.6.3
aniso8601==7.0.0
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
discord.py==1.5.1
Django==2.2.7
django-cors-headers==3.5.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
entrypoints==0.3
google==2.0.2
google-api-core==1.14.3
google-auth==1.7.0
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-storage==1.22.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==19.9.0
h11==0.9.0
h2==3.1.1
hpack==3.0.0
Hypercorn==0.8.2
hyperframe==5.2.0
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.2
jupyter-core==4.5.0
jupyterlab==1.1.3
jupyterlab-server==1.0.6
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
pipenv==2018.11.26
priority==1.3.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.9
protobuf==3.10.0
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.5
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.15.0
pytz==2019.2
pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
requests==2.22.0
rsa==4.0
Rx==1.6.1
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
tabulate==0.8.5
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
traitlets==4.3.2
typing-extensions==3.7.4
Unidecode==1.1.1
urllib3==1.25.6
virtualenv==16.7.5
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
wsproto==0.15.0
yarl==1.5.1
